A: Python - IPP #1 - see desc:
Moved problematic code from lines 30-39 into a comment section (""") for now

What this does is allows the program to run in the IDE. I'm not sure if the vscode python interpreter is playing up - exit().

Currently this just creates an endless stream of random names. This is because it is in the main function loop - while true. The names are generated, however there is no way to stop it or to "play again". The code that prevented the file from running in vscode's python interpreter is still there in comment form so I will put it into the default python IDE and hope that will work.

Github/acord-robotics
github/irisdroidology