A: Python - IPP #1 - see desc:.
So, on line 39 removed indent on the input for exiting, this allows for the program to run once (thus fixing the issue seen in the previous commit for this file). However, we then experience a traceback nameError, which is also seen in the screenshot attached to this repo and commit. Working through it.